+++ app/src/main/java/com/piwniczna/mojakancelaria/activities/clients/client_details/ClientDetailsFragment.kt
@@ -1,4 +1,4 @@
-package com.piwniczna.mojakancelaria.activities.client_details
+package com.piwniczna.mojakancelaria.activities.clients.client_details
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -9,9 +9,9 @@ import android.widget.Button
 import android.widget.TextView
 import com.piwniczna.mojakancelaria.Models.ClientEntity
 import com.piwniczna.mojakancelaria.R
-import com.piwniczna.mojakancelaria.activities.clients.ClientsFragment
+import com.piwniczna.mojakancelaria.activities.clients.clients_list.ClientsFragment
 import com.piwniczna.mojakancelaria.activities.clients.ObligationsFragment
-import com.piwniczna.mojakancelaria.activities.payments.PaymentsFragment
+import com.piwniczna.mojakancelaria.activities.payments.payments_list.PaymentsFragment
 
 class ClientDetailsFragment(var client: ClientEntity) : Fragment() {
     lateinit var titleTextView: TextView
@@ -36,7 +36,7 @@ class ClientDetailsFragment(var client: ClientEntity) : Fragment() {
     fun onBackPressed() {
         fragmentManager?.beginTransaction()?.replace(
             R.id.fragment_container,
-            ClientsFragment()
+                ClientsFragment()
         )?.commit()
     }
