+++ app/src/main/java/com/piwniczna/mojakancelaria/activities/obligations/obligations_list/ObligationsFragment.kt
@@ -2,7 +2,6 @@ package com.piwniczna.mojakancelaria.activities.clients
 
 import android.os.AsyncTask
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -16,13 +15,10 @@ import com.piwniczna.mojakancelaria.Models.ClientEntity
 import com.piwniczna.mojakancelaria.Models.ObligationEntity
 import com.piwniczna.mojakancelaria.Models.ObligationType
 import com.piwniczna.mojakancelaria.R
-import com.piwniczna.mojakancelaria.activities.add_client.AddClientFragment
-import com.piwniczna.mojakancelaria.activities.add_obligation.AddObligationFragment
-import com.piwniczna.mojakancelaria.activities.client_details.ClientDetailsFragment
-import com.piwniczna.mojakancelaria.activities.obligation_details.ObligationDetailsFragment
-import com.piwniczna.mojakancelaria.activities.obligations.ObligationsListAdapter
-import java.math.BigDecimal
-import java.time.LocalDate
+import com.piwniczna.mojakancelaria.activities.obligations.add_obligation.AddObligationFragment
+import com.piwniczna.mojakancelaria.activities.clients.client_details.ClientDetailsFragment
+import com.piwniczna.mojakancelaria.activities.obligations.obligation_details.ObligationDetailsFragment
+import com.piwniczna.mojakancelaria.activities.obligations.obligations_list.ObligationsListAdapter
 import kotlin.collections.ArrayList
